Append VoLTE status in the carrier name
for both expanded statusbar and keyguard.

Lately i've got a lot of requests for the bloody, hell, fu#kin, damn,
dieNow, justDie, ugly af, and crappy VoLTE icon.
I don't care about it but i hate it now :P

@xyyx has been too nice as usual and adapted caf code to show the icon
in the statusbar:
https://github.com/nitrogen-project/android_frameworks_base/commit/01bff2f43fcd254eb9bd01a09d9b7141a4d1af5a
https://github.com/nitrogen-project/android_packages_services_Telephony/commit/9593e3832950dbc97a30c06203442912a69bbccf

so we can use the same base to show a "VoLTE" flag in the carrier name
when it's available, using the existing CarrierText code to get the Ims state.
Also call straight the ITelephony aidl with Stub.
The services/telephony commit is still needed.

PS: i don't know if this works and i can't test the damn volte thing,
so if it doesn't work, meh.

Change-Id: I65b6a5109e4c9a959d150152c2a1e2cee100e59c