Work smarter not harder
The point of mirroring the analog clock class with the 'Custom Analog Clock'
class was to avoid any issues with any apps or whatever. With the dead pool
analog clock we don't need to do this. We can extend to the 'Custom Analog
Clock' class and avoid repetitive code calls.

Change-Id: I62cb0d919069f405ef690e3a9e5d85877c29f7ca
Signed-off-by: RakeshBatra <raakesh.batra@rediffmail.com>